+++ archive/validator/validate.py
@@ -55,7 +55,7 @@ import logging
 
 from acetime.zonedb_types import ZoneInfoMap
 from acetime.zonedb_types import ZonePolicyMap
-from acetimetools.data_types.at_types import TransformerResult
+from acetimetools.datatypes.attyping import TransformerResult
 from acetimetools.extractor.extractor import Extractor
 from acetimetools.transformer.transformer import Transformer
 from acetimetools.zone_processor.zone_info_inliner import ZoneInfoInliner