+++ src/acetimetools/generator/cgenerator.py
@@ -10,9 +10,9 @@ import os
 import logging
 from typing import List
 
-from acetimetools.data_types.at_types import ZoneRuleRaw
-from acetimetools.data_types.at_types import ZoneEraRaw
-from acetimetools.data_types.at_types import ZoneInfoDatabase
+from acetimetools.datatypes.attyping import ZoneRuleRaw
+from acetimetools.datatypes.attyping import ZoneEraRaw
+from acetimetools.datatypes.attyping import ZoneInfoDatabase
 from acetimetools.transformer.transformer import normalize_name
 from acetimetools.transformer.transformer import normalize_raw
 from acetimetools.generator.argenerator import compressed_name_to_c_string